% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemcalc.R
\name{calc_molarity}
\alias{calc_molarity}
\title{Molarity Calculator}
\usage{
calc_molarity(moles, volume_L)
}
\arguments{
\item{moles}{The amount of solute in moles.}

\item{volume_L}{The volume of the solution in liters.}
}
\value{
The molarity of the solution as a numeric value.
}
\description{
This function calculates the molarity (M) of a solution, given the number of moles of solute and the volume of the solution in liters.
}
\examples{
calc_molarity(moles = 0.5, volume_L = 1)
}
