% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elements}
\alias{elements}
\title{Elements Dataset}
\format{
A data frame with 118 rows and 3 columns:
\describe{
\item{Symbol}{Chemical symbol of the element.}
\item{Name}{Full name of the element.}
\item{AtomicMass}{Atomic mass of the element.}}
}
\source{
https://pubchem.ncbi.nlm.nih.gov/ptable/atomic-mass/.
}
\usage{
data(elements)
}
\description{
A dataset containing information about chemical elements, including their
symbol, name, and atomic mass.
}
\keyword{datasets}
