% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemcalc.R
\name{percent_yield}
\alias{percent_yield}
\title{Percent Yield Calculator}
\usage{
percent_yield(actual, theoretical)
}
\arguments{
\item{actual}{The actual yield obtained from the experiment.}

\item{theoretical}{The theoretical yield based on stoichiometric calculations.}
}
\value{
The percent yield as a numeric value.
}
\description{
This function calculates the percent yield of a chemical reaction.
}
\examples{
percent_yield(actual = 3.5, theoretical = 5)
}
